sound: sunxi acx00 - support for speaker state PM
This change allows the external speaker gpio to be enabled
or disabled after a resume-from-sleep event based on
the state of the speaker GPIO prior the suspend event.
Previosuly the speaker switch was set to on-state, always.
Now, the speaker can be on or off (set via userspace)
and after suspend/resume the speaker switch value is preserved.