driver: ssv6x5x: removed SSV6X5X_PM_MOD functionality
Previously, the PM_MOD was used to enable suspend/resume functionality
which was (seemingly) not working. The results of the PM_MOD were
not stable. The driver can suspend/resume without the PM_MOD
if the 'wowlan' feature is enabled from user space. It is possible to
enable the feature either by running the following command:
iw phy0 wowlan enable magic-packet
or by setting the following option in wpa_supplicant:
wowlan_triggers=any